Backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
google-generativeai==0.x.x